Repurpose recursivelyWalkCallTargetAndPerformAction
This routine was only being used for recursively walking call
targets and counting nodes, despite being designed to perform
any "action". It was not going to work for any action other
than generateNodeEstimate as the debugTrace message relied
on calling generateNodeEstimate::getNodeEstimate.

This commit repurposes this function to only be used for node
counting, rather than any "action", and renames the function
accordingly.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>