Displacement API for offheap using region flags
Improve API for displacement calculation of stack referenced arrays for
offheap to rely on region flags to determine the type of the movement
(evacuate vs sliding). Previously the movement type was determined by
RootScanner. But unlike standard platform threads, Virtual Threads
(Continuations) are not roots and their stack walker does not
instantiate RootScanner.

Relying on region flags (of source object) works for both types of
threads.

Signed-off-by: Aleksandar Micic <Aleksandar_Micic@ca.ibm.com>