Ensure all JITServer compilation threads process CH Table Updates
Under very rare circumstances, it is possible for a JITServer
compilation thread processing a remote compilation request containing CH
Table updates to side step processing the CH Table updates. The
sequences of events that could lead to this is as follows:

1. compThreadID=0 processes request seqNo=2 (isCritical=0) (criticalSeqNo=1 lastProcessedCriticalReq=1)
2. compThreadID=1 processes request seqNo=1 (isCritical=1) (criticalSeqNo=0 lastProcessedCriticalReq=1)
3. compThreadID=1 discards request because seqNo=2 was processed first
4. compThreadID=0 requests full CH Table to initialize on the server
5. compThreadID=1 processes request seqNo=3 (isCritical=1) (criticalSeqNo=1 lastProcessedCriticalReq=1)

seqNo=1 and seqNo=3 contain CH Table updates; seqNo=2 does not. As such,
on the client when generating seqNo=3, it still only depends on seqNo=1.
This means that on the server, by step 4 above, both compThreadID=0 and
compThread=1 can reach the point where the code determines whether or
not to initialize the CH Table.

Until compThreadID=0 finishes initializing the CH Table, the
clientSession->cachesAreCleared() flag remains true. This causes
compThreadID=1 to block on the monitor guarding the initialization code.
Once compThreadID=0 is done initializing the CH Table, compThreadID=1 on
longer sees clientSession->cachesAreCleared() as true (after it
unblocks).

However, the way the code is structured, this results in compThreadID=1
not executing any code that processes the CH Table updates associated
with its compilation request.

This commit fixes this by restructuring the code to ensure that all
threads either execute the CH Table initialization code or the CH Table
updates code.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>