Add -Xmx512m to -Xcheck:memory GC Regression tests
Looks like -Xcheck:memory has a problem on x86 Mac platform preventing
to run Balanced with enabled off-heap with default maximum heap size as
25% of RAM available on machine failing with Native OOM. I guess it
forces Sparce Heap to be allocated in the physical memory instead of be
mostly virtual. Size of the heap is not significant for this particular
test, so it can be explicitly set to smaller value to avoid the problem.
We are going to investigate -Xcheck:memory behaviour on x86 Mac.