JEP491: Never Deflate Monitors and Synchronize virtualThreadWaitCount
Currently, there are timing holes between
JVM_TakeVirtualThreadListToUnblock and monitor deflation. A monitor can
be deflated while it is being accessed in
JVM_TakeVirtualThreadListToUnblock. This leads to a NULL dereference
causing a segfault. Adding more synchronization will cause a
significant overhead in the object monitor exit path. Until an
efficient solution is developed, the policy to never deflate will be
employed in order to support JEP491. Since the current JEP491
implementation always inflates monitors before usage, deflating will be
counter-productive.

Also, added a null check for syncObjectMonitor in
JVM_TakeVirtualThreadListToUnblock to ensure that the monitor is
inflated before operations are performed on it.

Operations on J9ObjectMonitor's virtualThreadWaitCount field may
happen out of sequence. To ensure correct ordering and consistency,
atomics have been employed to modify this field.

Related: #20705

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>