Propagate known key info to ConcurrentHashMap.get
ConcurrentHashMap.get is similar in nature to HashMap.get operations,
and we can benefit from compile-time folding of various queries such
as Object.hashCode() if the object being passed as arg is known and
fixed. For that to be possible, we need to ensure that the caller
method of ConcurrentHashMap.get() undergoes peeking IL-gen so that
the object information can be passed down and used.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>