Ensure Escape Analysis tracing prints candidate node
Some calls to traceMsg print the address of the Candidate table entry,
rather than the candidate node - the _node field in the Candidate
structure - itself.  Cleaned up those remaining calls, including some
that were commented out, by ensuring the _node field is printed.

Also added a test of trace() for a call to traceMsg() that was
previously called unconditionally.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>