Remove frame build in interpreter monenter
The interpreter stack is walkable on the monitor enter bytecode so no
frame build is required. Additionally, the internal native frame build
is incorrect as it assumes the number of pending pushes is the number of
args to the method.

Object.wait still needs a frame build since its an INL.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>