Fix frame build in vthread transistions
The interpreter stack is walkable on the monitor enter bytecode so no
frame build is required. Additionally, the internal native frame build
is incorrect in this case as it assumes the number of pending pushes is
the number of args to the method.

Object::wait and Continuation::yield still need a frame build since its
an INL.

Patch the J9VMTHREAD register upon stack swap so JIT transition back to
JIT uses the current vmthread.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>