JITServer support for ssl libs name changes
When encryption is enabled for JITServer, the server
will attempt to load the crypto and ssl libraries.
The server maintains a list of library names (ordered
by priority). In the near future OpenJ9 JVM distributions
will embed crypto/ssl libraries with specific names.
This commit adds those specific names to the list
of libraries that the server will be looking for.

Signed-off-by: Marius <mpirvu@ca.ibm.com>