Fix TR_ResolvedJ9Method::isSameMethod() for native methods
The prior logic here would refuse to acknowledge that two instances of
TR_ResolvedJ9Method represent the same method if that method was native.
A single instance wouldn't even be treated as the same method as itself.
For most purposes, that negative result is incorrect and surprising.

The reason for excluding native methods was that when compiling a
linkage adapter for a JNI method, the call would otherwise appear to be
recursive according to isRecursiveMethodTarget(), which could cause code
generators to encode the call instruction as a call to the JIT body
instead of the native function that implements the method. This is now
handled in isRecursiveMethodTarget() so that isSameMethod() can return
the obviously correct result.