Add support for cached ffi_type object for z/OS
In https://github.com/eclipse-openj9/openj9/pull/19988, changes were added to
cache the ffi_type for same struct in downcall. LIBFFI on z/OS to handle the
complex type arguments (An aggregate containning exactly two floating point
values of same type) would look at the shape of the structure and change the
type so that HLASM code that acts as an interface to downcall can arrange
arguments correctly. With caching of the ffi_type, we were missing setting the
flag for return type if we encounter the cached ffi_type object causing us to
miss out setting up return values.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>