Add options to enable the JITServer profile cache
This commit adds two external options that enable/disable the
JITServer profile cache from the client point of view:
`-XX:+JITServerUseProfileCache` and `-XX:-JITServerUseProfileCache`
The JITServer profile cache mechanism is always enabled at the server.
A client merely indicates its intent to use the facility the server
provides. Note that the enablement of the JITServer AOT cache is a
pre-requisite for the JITServer profile cache.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>