Updates to JITServerSharedROMClassCache
This commit implements the following changes:

- Adds a public routine that increments the reference count of
a ROMClass stored in the cache. This is used to prevent the
deletion of the ROMClass while it is still in use.

- Adds a routine that prints to stderr the entire content of
the JITServerSharedROMClassCache for debugging purposes.

- Modifies the `JITServerSharedROMClassCache::shutdown(bool lastClient)`
routine so that it does not delete the entire cache when the
last client disconnects. The JITServer shared profile cache may
refer to ROMClasses in JITServerSharedROMClassCache even if no
clients are currently attached.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>