Add #if defined(J9VM_OPT_SNAPSHOTS) around Snapshot utility macros
As per https://github.com/eclipse-openj9/openj9/pull/21667#discussion_r2049086579
When the feature is disabled, IS_RESTORE_RUN is always FALSE. The
previous assumption was that the compiler will optimize the code block
and evaluate it as dead code, removing it. However, we've observed perf
issues on zLinux where the compiler fails to optimize correctly. Just to
be safe, it should be #ifdef'ed.

Signed-off-by: Jason Feng <fengj@ca.ibm.com>