Refine and inline MethodHandle.linkTo* methods in ValuePropagation
When ValuePropagation is able to obtain the known object info of
the MemberName child of a linkToSpecial, linkToVirtual and
linkToStatic call, it is possible to refine the INL call and avoid
a more expensive dispatch mechanism. The refined linkTo* INL call
then gets inlined during delayed VP transformations.

This change also adds a new helper to deal with populating
prex arg info and adding the CallInfo to the refined INL
call list.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>