Add table of invokeBasic() call sites to J9JITExceptionTable
It is possible for MethodHandle.linkToVirtual() to be called with a
MemberName that specifies MethodHandle.invokeBasic() as the virtual
method. For performance, any linkToVirtual() call that goes unrefined in
a JIT compilation will be transformed into a computed call through the
vtable using JITHelpers.dispatchVirtual(). When the target happens at
runtime to be invokeBasic(), this will result in a call through the
vtable entry for invokeBasic(), which has been the cause of multiple
problems that are fixed by this commit.

1. old_slow_icallVMprJavaSendPatchupVirtual() would try to find the J2I
   thunk based on the signature from the ROM method, which doesn't
   necessarily (or even probably) match the signature at the call site
   because invokeBasic() is signature-polymorphic. The J2I thunk might
   not exist, in which case the lookup would return null, and execution
   would jump to address 0. Even if it did exist, the arguments likely
   wouldn't be copied correctly to the stack. There needs to be some way
   for the correct J2I thunk to be determined.

2. old_slow_icallVMprJavaSendPatchupVirtual() would update the vtable
   entry to point to the J2I thunk. Even if it had the correct J2I thunk
   for the particular call that was underway at the time, later calls
   from different call sites might require a different J2I thunk (again
   due to signature-polymorphism). For invokeBasic(), the vtable entry
   must be left as-is so that the thunk can be determined separately for
   each call.

3. For J2I invokeBasic(), the VM expected J9VMThread.tempSlot to specify
   the number of stack slots for the arguments. This was not set by the
   JIT prior to dispatchVirtual() calls, but even if it were, the count
   would be wrong when calling invokeBasic() that way because tempSlot
   would be clobbered before reaching the interpreter. Specifically,
   old_slow_icallVMprJavaSendPatchupVirtual() receives the JIT vtable
   offset (for cases where it isn't immediate) in tempSlot, and then it
   uses tempSlot again to specify the thunk to which control should be
   transferred.

Now, for any JIT body containing a call to either invokeBasic() or
dispatchVirtual(), the J9JITExceptionTable has a table of all such call
sites, identified by the offset from startPC to the JIT return address.
Each entry specifies the number of stack slots for the arguments, and in
the case of dispatchVirtual(), the J2I thunk. This allows the necessary
information to be looked up based solely on the JIT return address both
in old_slow_icallVMprJavaSendPatchupVirtual() and in the interpreter's
invokeBasic().

(cherry picked from commit 897d9f9a2eb3d6d9a83deab823ee986e8c491a64)