Add tests for BSM resolution on error throw
When an error is thrown during resolveInvokeDynamic, we return a special
error-throwing MethodHandle. This MH has no paramters, which can cause
problems for OSR Bookkeeping since no arguments will be popped off the
stack, causing the assertion failure in ILGen "cannot create a pending
push temporary that exceeds that number of slots for this method" since
we are trying to push the return value onto the stack when it is already
full

This commit tests arguments are popped and that the assertion does not
occur.

Signed-off-by: Matthew Hall <matthew.hall3@outlook.com>