Anchor the value node before the removal of the call node
Need to anchor the value node before the helper call node is removed.
Otherwise, the child value node could be currently anchored under the
helper call node. When the helper call node is removed, the value node
will be moved down and anchored where the next reference is.
It will be a problem if there is a store into this value between the
helper call node and the next reference. After the helper call node
is removed, the reference will load the updated value instead of
the original value.

Fixes: #21256
Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>