Raise allowance size threshold along alwaysWorthInlining calls
In Estimate Code Size, we would terminate estimation once
_analyzedSize exceeded the analyzed size threshold, which was
obtained using sizeThreshold * allowanceFactor. However, not
increasing the analyzed size threshold for methods that are always
worth inlining - a property set through inliner policy or through
the use of @ForceInline annotation, meant that those methods were
not being prioritized as necessary. To overcome that issue, this
commit introduces a few changes to the ECS heuristics:
  * forceInlineMultiplier - to be used to multiply the analyzed
    size threshold for a method that is always worth inlining.
  * Change estimateCodeSize function signature to be able to
    recursively pass down caller method's analyzed size threshold
    to callees along an always worth inlining call graph.
  * Reset _analyzedSize to pre-analysis _analyzedSize for methods
    always worth inlining.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>