Fix Class Loading gap during relocation
The relocation infrastructure did not acquire the Class Table Mutex at
any point. As a consequence, it was possible for an inlined method to
get overridden between the point where the validation occurs (in
preparePrivateData), and the point where the assumptions are
registered (in applyRelocationAtAllOffsets).

This commit fixes this by acquring the Class Table Mutex in
TR_RelocationRecordGroup::handleRelocation, which ensures that it is
held across both the validation and the registration of the assumptions.
Then, to ensure functional correctness with the SVM, in
TR_RelocationRecordInlinedMethod::inlinedSiteValid,
virtualMethodIsOverridden is used to ensure that the method has not been
overridden since the method was materialized during validation.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>