Update IntegerToChars evaluators for off-heap
The acceleration was initially disabled because it added array header
size to the array object, which was required to reach the elements.
However, this is unnecessary for off-heap mode. As a result, the
evaluator has been updated to add the array header size exclusively
for non-off-heap mode, while in off-heap mode, it now retrieves the
data element address directly from the array header, enabling
acceleration in off-heap mode.

Signed-off-by: Shubham Verma <shubhamv.sv@gmail.com>