Refactor wasDeserialized flag
The wasDeserialized flag is used to inform other threads that the
deserializer was reset. This used to be a member of the
TR_CompilationInfoPerThread class. However, this class is only
guaranteed to exist for compilation threads. In preparation to support
AOT Method Dependencies for JITServer, a non-compilation thread may need
to invoke the deserializer. As such, there needs to be a way to inform
non-compilation threads that the deserializer was reset.

This commit refactors out the flag and places it in the TR_J9VMBase
class. It also adds an API to allow non-compilation threads to register
with the Deserializer to be notified about when the deserializer gets
reset.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>