Create TR_IPBCDataDirectCall for IProfiler
The IProfiler does not collect samples for direct calls.
Instead, the optimizer will use the setCallCount() API to
set the number of samples for such direct calls.
Until this commit, direct-call info was stored in
TR_IPBCDataCallGraph entries which was mainly created for
interface and virtual calls. Such entries track 3 different
call targets and are somewhat bigger than the rest of IProfiler
entries. For direct calls we only need a simple counter, so
using TR_IPBCDataCallGraph entries for this purpose is wasteful.

This commit introduces a new type of IProfiler entry, called
TR_IPBCDataDirectCall, which will be used for storing profiling
info related to direct calls. Note that the optimizer will
remain the sole entity that will populate such entries.
Moreover, these new entries will not be stored into the SCC
(this follows the existing approach where profiling info related
to direct calls is not stored in SCC).

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>