Fix is{Anonymous,Hidden}Class() for 64-bit big-endian JITServer
For CLASSINFO_ROMCLASS_EXTRAMODIFIERS the result is uint32_t, but these
call sites were passing a pointer to uintptr_t, which on 64-bit is twice
as big as necessary. On a big-endian system, the result would be stored
to the high half of the variable, leaving the low half zero, so these
functions would return false for all classes.