Enable peeking ILGen for param loads using NeedsPeekingHeuristic
NPH keeps track of loads of parameters when the caller method has
prex arg info, which is then used when we come across one of the
invoke* bytecodes. If the param load is used as arg for the call,
then we would want to propagate the prex arg info down the
callee, for which peeking ILGen is necessary. This commit also
terminates NPH once _needsPeeking is set as true, as any further
analysis would not change the peeking outcome. Peeking is disabled
by default for MethodHandle thunk archetype methods in JDK8 and
JDK11, as they are not typical Java methods and require special
handling in existing optimizations to refine and inline.

Peeking ILGen due to param loads can be disabled using the new env
option TR_disableNPH.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>