JITServer: Recognize cross-loader superclasses in isSameOrSuperClass()
It's possible for a class to be defined by a different loader than its
superclass, in which case this method would incorrectly return false on
JITServer. Now it will ignore class loaders like the base implementation
does.

So far the incorrect result would only have caused the SVM to fail to
notice that a class is a subclass of Throwable if the class was not
defined by the bootstrap loader. Then it would remember such a class
even though it's supposed to be considered not to be worth remembering.