Use the actual declaring class for static final field folding
Static final field folding has been looking at the class named in the
constant pool entry instead of the actual declaring class. These are
often but not always the same class. When they differ, the declaring
class is a supertype of the named class.

This could allow static final fields to be folded even if they are
declared by a class that has seen illegal stores, or that contains an
illegal putstatic instruction outside of <clinit>, or that is supposed
to be excluded based on dontFoldStaticFinalFields={...}. For guarded
folding, it could also cause the runtime assumption to be registered for
the subtype instead of the declaring class, in which case the folded
field could be modified later without invalidating the assumption.