Enable bigCallee size adjustment during ECS for JDK8 and 11
During EstimateCodeSize, if we came across a callee method that
was deemed too big, and therefore, set to be discarded as an inlining
candidate, a size adjustment for other methods above such methods is
necessary to ensure the entire call graph does not get marked as
too big. For JDK8 and 11, this adjustment was disabled due to
a functional issue that was exposed due to the different inlining
outcome from this adjustment. This change removes the restriction
that was in place as a workaround, and fixes performance issues
resulting from early ECS termination in JDK8 and 11.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>