Enhance visitStoreToLocalVariable in MethodHandleTransformer
visitStoreToLocalVariable can now be triggered for cases of direct
stores of nodes anchored by PassThrough nodes, which would result
from calls that have been transformed into PassThrough. PassThrough
nodes do not have type info, and therefore this has to be accounted
for. Once inside visitStoreToLocalVariable, it would have to
recognize cases of stores of child nodes of PassThrough, and set the
actual node to use as the RHS of the store.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>