Defer generating dereference of vft-symbol in JProfilingValue
In the process of finding values to profile, the method
TR_JProfilingValue::cleanUpAndAddProfilingCandidates would insert calls
to the jProfileValueWithNullCHK non-helper for instanceof and checkcast
operations, with the child of each call being an indirect load of the
<vft-symbol> for the value that is to be profiled.

During TR_JProfilingValue::addProfilingTrees, those calls would be
transformed to guard the dereference of the indirect load with a test of
whether the value is a null reference.

However, if generation of profiling trees is halted early during
TR_JProfilingValue::lowerCalls, any value beneath the
jProfileValueWithNullCHK will simply be left anchored in its original
position, unguarded.  This can result in a segmentation fault if the
value that was to be profiled was a null reference.

This change delays introducing the indirect load of the <vft-symbol>
until TR_JProfilingValue::lowerCalls, so that the indirect load is
properly guarded.  If generation of profiling trees is halted early, the
original value will be left behind anchored, allowing it be properly
evaluated in its original relative position.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>