setWarmCallGraphTooBig() creates IProfiler entry if needed
setWarmCallGraphTooBig is an IProfiler routine used by the
Inliner as a hint that the call at a particular callsite
is expensive to be inlined.
Static and special invokes are not tracked by the IProfiler
and therefore profiling entries for such calls are not created,
yet we may want to set the warmCallTooBig hint for direct calls
as well. We have empirical evidence that not doing so increases
the compilation overhead by ~5%.
Having said that, the Inliner indirectly creates profiling entries
for direct calls when using setCallCount(), so, most of the time
the inliner can set the warmCallTooBig hint if it wants to.
However, we plan to eliminate the setCallCount()/getCallCount()
pair of routines and, to compensate, this PR creates IProfiler
entries for static/special invokes in the setWarmCallGraphTooBig()
routine.

Signed-off-by: Marius <mpirvu@ca.ibm.com>