Add Setup stage between Queue and Get Source
Once the machine is acquired, there are several things happening before
we hit the get_source script. This is currently being displayed in the
Queue stage which is misleading. Adding a Setup stage here to separate
these parts of the build out to make it more clear where the time is
being spent.

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>