Update class loader graph for review (will squash)
Additionally, identify the permanent loaders by getting them from
systemClassLoader, extensionClassLoader, and applicationClassLoader, but
only once they're fully initialized and frozen. It's possible for the
latter two to be replaced during bootstrapping.

Permanent loaders don't need an outliving loader set, since they're
outlived only by other permanent loaders, so instead outlivingLoaders is
set to a special value to indicate that the loader is permanent.

Permanence is propagated in that any loader that is found to outlive a
permanent loader must also be permanent. This will be taken into account
regardless of which is known first: the permanence or the lifetime
relationship.

Permanent loaders are reported to the JIT so that it can take advantage
of them when reasoning about class loader lifetimes. It is now possible
to get a vector of pointers to all permanent loaders known so far by
calling J9::Compilation::permanentLoaders().

At the moment, outlivingLoaders and permanentLoaders() are essentially
unused, but soon they will both be used for RetainedMethodSet.