Add indirect load of vft-symbol to quick test
If the operation that's being profiled is instanceof or checkcast, the
code in the quickTest block and the helper call block should test the
result of the load of the vft-symbol for profiling.  The previous change
only added the indirect load of the vft-symbol in the IL generated for
the helper call.  This change adds the missing load of the vft-symbol to
the quickTest block as well.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>