RCP: Introducing Frozen Classes
During snapshot creation, the J9Class structures whose class loader is
application class loader are marked as frozen when their classObject
fields are nullified. During the restore phase, these structures are never
used, kept hidden from the GC and other external users as a limitation.
This is achieved by updating the skip logic in the J9HashTable,
J9MemorySegment Iterator, which prevents the GC from encountering
J9Class structures in an unusable state that could lead to crashes.
The class ClassHeapIterator is refactored to use a generic vm interface
segmentIteratorNextClass() to iterate over classes in a segment, which
implements the codes in the case of J9VM_OPT_FRAGMENT_RAM_CLASSES defined
only.

Fixes: #21892

Co-authored-by: Babneet Singh sbabneet@ca.ibm.com
Co-authored-by: Tobi Ajila tobi_ajila@ca.ibm.com
Co-authored-by: Lige Zhou lige.zhou@ibm.com