Eliminate VM_compiledAsDLTBefore message for JITServer
Recent changes in Inliner have added a check of whether a method
has been DLTed before (DLT stands for Dynamic Loop Transfer).
As a consequence, the number of `VM_compiledAsDLTBefore` messages
sent by the server has increased a lot adding to compilation overhead.

This commit eliminates all `VM_compiledAsDLTBefore` messages.
The idea is for the server to keep track of all DLT compilations
that it performs on behalf of a client. When the client connects
for the first time to a new server, it will send the set of DLTed
methods to the server, so that the server can initialize its own set.
This action takes care of two possible situations:
(1) The client connects to a server at a later point in time (because
a server was not available from the very begining).
(2) The client is connected to a server at the begining of its run,
but later on, the connection is broken and the client is forced to
connect to a different server.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>