Ensure method symbol is a resolved method before requesting it
In J9::Compilation::getReloTypeForMethodToBeInlined, for the Abstract
Method scenario, the compiler needs the resolved method from the call
node's method symbol to determine whether the called method is abstract.
However, interface calls are treated as unresolved. This can cause a
crash.

Fix this by first checking if the method symbol is a resolved method
before acquiring it.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>