Use a "defaultMonitorSpinCount1" value of 1 when running on a single CPU
When running on a single CPU, a tight spin loop may only waste CPU
resources because there the thread that holds the lock cannot run
on another CPU and release the lock. In such a case it's better to
directly yield the CPU to another thread, hopefully the thread
that holds the lock.

Signed-off-by: Marius <mpirvu@ca.ibm.com>