Report optimization analysis phase with -Xjit:vmState option
In the event of a compilation crash, a VM state might be reported that
is not recognized by the -Xjit:vmState option.  This is most likely to
happen if the crash occurs following an optimization while performing
checks of the IL, CFG or structure, but could happen during other
analysis phases associated with an optimization that are not directly a
part of the optimization itself.  The Options::vmStateOption method was
written assuming that the low byte of the VM State will always be 0xFF
if the crash happened during the optimization phase, but that is only
true if the crash happened in the midst of a particular optimization.

This change takes advantage of the new
OMR::Optimizer::getAnalysisPhaseName method to get the name of other
optimization phases that might be encoded in the VMState.  It also
updates documentation to reflect the fact that the low byte need not
have the value 0xFF if a crash occurs during some analysis phase
associated with optimization, rather than in the optimization itself.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>