Fix crash under -Xaot:
When running with -Xjit: (i.e., not -Xjit but -Xjit: which has the
colon followed by an empty string), the JVM prints out an error
message and terminates. However, when running with -Xaot: the error
message is printed out but then the JVM crashes. This happens because
there is a call to codert_onload between initializing the compiler args
for -Xjit and -Xaot.

This commit fixes this by invoking codert_onload after initializing both
-Xjit and -Xaot. This is ok because the initization does not actually
process the options. Rather, it sets the value of a pointer to point to
the compiler args. Thus, calling codert_onload after initializing both
is safe because the initialization does not depend on any initialization
that codert_onload performs.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>